+++ mvc/controllers/userCtrl.php
@@ -1,41 +1,107 @@
 <?php
 
 	
-	class UserCtrl extends ValidationCtrl{
+	class UserCtrl extends StandardCtrl{
 		private $model;
 	  
 		function __construct(){
+			parent::__construct();
 	    	require('models/userMdl.php');
 	    	$this->model=new UserMdl();
 		}
 	  
 		function run(){
 			switch($_GET['act']){
+				case 'login':
+					if($this->isLogged()==FALSE){
+						if($this->login()){
+							require('views/successLogin.php');
+						}
+					}
+					break;
+				case 'logout':
+					if($this->isLogged()==TRUE){
+						$this->logout();
+						require('views/logout.php');
+					}
+					break;
 				case 'create':
 		  			//User is valid and have permissions
-		  			$this->create();
+		  			if($this->isLogged()){
+		  				if($this->isManager()){
+		  					$this->create();
+		  				}
+						else{
+							require('views/NoAccess.php');
+						}
+		  			}
+					else{
+						require('views/needLogin.php');
+					}
 		  			break;
 				case 'delete':
 					//User is valid and have permissions
-					$this->delete();
+					if($this->isLogged()){
+		  				if($this->isManager()){
+		  					$this->delete();
+		  				}
+						else{
+							require('views/NoAccess.php');
+						}
+		  			}
+					else{
+						require('views/needLogin.php');
+					}
 					break;
 				case 'update':
-					$this->update();
 					//User is valid and have permissions
+					if($this->isLogged()){
+		  				if($this->isManager()){
+		  					$this->update();
+		  				}
+						else{
+							require('views/NoAccess.php');
+						}
+		  			}
+					else{
+						require('views/needLogin.php');
+					}
 					break;
 				case 'select':
 					//User is valid and have permissions
-					$this->select();
+					if($this->isLogged()){
+		  				if($this->isManager() or $this->isUser()){
+		  					$this->select();
+		  				}
+						else{
+							require('views/NoAccess.php');
+						}
+		  			}
+					else{
+						require('views/needLogin.php');
+					}
 					break;
 				case 'list':
 					//User is valid and have permissions
-					$this->listUsers();
+					if($this->isLogged()){
+						
+		  				if($this->isManager() or $this->isUser()){
+		  					$this->listUsers();
+		  				}
+						else{
+							require('views/NoAccess.php');
+						}
+		  			}
+					else{
+						require('views/needLogin.php');
+					}
 					break;
 				default:
 				  break;
 			}
 		}
 	
+		
 		/**
 		 * Create a new user. Doing validation over input data
 		 */
@@ -43,6 +109,7 @@
 			$Correct=TRUE;//Flag to determine if it can create a new user
 			$NoSet=FALSE; //Flag to determine if the variables are set
 	  		//Validate variables and if variables is set 
+			$User=isset($_POST['user'])?$this->validateUserName($_POST['user']):$NoSet=TRUE;
 			$Name=isset($_POST['Name'])?$this->validateName($_POST['Name']):$NoSet=TRUE;
 			$MaternalLastname=isset($_POST['MaternalLastname'])?$this->validateName($_POST['MaternalLastname']):$NoSet=TRUE;
 			$PaternalLastname=isset($_POST['PaternalLastname'])?$this->validateName($_POST['PaternalLastname']):$NoSet=TRUE;
@@ -102,11 +169,11 @@
 				//Insert the new User
 				if($Correct==TRUE){
 					
-					$Result=$this->model->create($Name,$MaternalLastname,$PaternalLastname,$Email,$Job,$Telephone);     
+					$Result=$this->model->create($User,$Name,$MaternalLastname,$PaternalLastname,$Email,$Job,$Telephone,$Pass);     
 					if($Result==TRUE){
 						//Shown a view
 						require('views/userInserted.php');
-						require('mail.php');
+						require('controllers/mail.php');
 						$subject = 'Correo de registro de usuario';
 						$body = 'El usuario ' . $Name . ' ' . $MaternalLastname . ' ' . $PaternalLastname . ' se ha registrado correctamente';
 						$mail = new Email($Email, $subject, $body);
@@ -199,7 +266,7 @@
 					$Result=$this->model->update($ID,$Name,$MaternalLastname,$PaternalLastname,$Email,$Job,$Telephone);
 		 			if($Result){
 		  				require('views/userModified.php');
-		  				require('mail.php');
+		  				require('controllers/mail.php');
 						$subject = 'Correo de actualizacion de datos usuario';
 						$body = 'Los datos del usuario ' . $Name . ' ' . $MaternalLastname . ' ' . $PaternalLastname . ' se han actualizado.';
 						$body = $body . ' Email: ' . $Email;
